@@ -1,38 +1,35 @@
 <?php
 namespace PhpAbac\Test;
 
-date_default_timezone_set('UTC');
-
-
 use PhpAbac\Abac;
 
-class AbacTest extends \PHPUnit_Framework_TestCase
+class AbacTest extends \PHPUnit\Framework\TestCase
 {
-    /** @var Abac[] $abac_a * */
-    protected $abac_a;
-    /** @var Abac[] $abac_array_a * */
-    protected $abac_array_a;
-    /** @var Abac[] $abac_getter_params_a * */
-    protected $abac_getter_params_a;
-    /** @var Abac[] $abac_import_a * */
-    protected $abac_import_a;
+    /** @var array * */
+    protected $bsicSet;
+    /** @var array * */
+    protected $multipleRulesetSet;
+    /** @var array **/
+    protected $getterParamsSet;
+    /** @var array **/
+    protected $importSet;
 
     public function setUp()
     {
-        $this->abac_a = [
+        $this->basicSet = [
             new Abac([__DIR__ . '/fixtures/policy_rules.yml']),
             new Abac([__DIR__ . '/fixtures/policy_rules.json']),
         ];
-        $this->abac_array_a = [
+        $this->multipleRulesetSet = [
             new Abac([__DIR__ . '/fixtures/policy_rules_with_array.yml']),
             new Abac([__DIR__ . '/fixtures/policy_rules_with_array.json']),
             new Abac(['policy_rules_with_array.yml'], [], __DIR__.'/fixtures/'),
             new Abac(['policy_rules_with_array.json'], [], __DIR__.'/fixtures/'),
         ];
-        $this->abac_getter_params_a = [
+        $this->getterParamsSet = [
             new Abac(['policy_rules_with_getter_params.yml'], [], __DIR__.'/fixtures/'),
         ];
-        $this->abac_import_a = [
+        $this->importSet = [
             new Abac(['policy_rules_with_import.yml'], [], __DIR__.'/fixtures/'),
         ];
     }
@@ -44,30 +41,23 @@ class AbacTest extends \PHPUnit_Framework_TestCase
         $users = include('tests/fixtures/users.php');
         $vehicles = include('tests/fixtures/vehicles.php');
 
-        foreach ($this->abac_a as $abac) {
+        foreach ($this->basicSet as $abac) {
             $this->assertTrue($abac->enforce('nationality-access', $users[3]));
-            $this->assertEquals([
-                'japd',
-            ], $abac->enforce('nationality-access', $users[1]));
+            $this->assertFalse($abac->enforce('nationality-access', $users[1]));
+            $this->assertEquals(['japd'], $abac->getErrors());
 
             // getenv() don't work in CLI scripts without putenv()
             putenv('SERVICE_STATE=OPEN');
 
             $this->assertTrue($abac->enforce('vehicle-homologation', $users[0], $vehicles[0]));
-            $this->assertEquals(
-                ['derniere-revision-technique'],
-                $abac->enforce('vehicle-homologation', $users[2], $vehicles[1])
-            );
-            $this->assertEquals(
-                ['permis-de-conduire'],
-                $abac->enforce('vehicle-homologation', $users[3], $vehicles[3])
-            );
-            $this->assertEquals(
-                ['visas'],
-                $abac->enforce('travel-to-foreign-country', $users[0], null, [
-                    'dynamic_attributes' => ['code-pays' => 'US']
-                ])
-            );
+            $this->assertFalse($abac->enforce('vehicle-homologation', $users[2], $vehicles[1]));
+            $this->assertEquals(['derniere-revision-technique'], $abac->getErrors());
+            $this->assertFalse($abac->enforce('vehicle-homologation', $users[3], $vehicles[3]));
+            $this->assertEquals(['permis-de-conduire'], $abac->getErrors());
+            $this->assertFalse($abac->enforce('travel-to-foreign-country', $users[0], null, [
+                'dynamic_attributes' => ['code-pays' => 'US']
+            ]));
+            $this->assertEquals(['visas'], $abac->getErrors());
             $this->assertTrue($abac->enforce('travel-to-foreign-country', $users[1], null, [
                 'dynamic_attributes' => ['code-pays' => 'US']
             ]));
@@ -75,16 +65,16 @@ class AbacTest extends \PHPUnit_Framework_TestCase
     }
 
 
-    public function testEnforceArray()
+    public function testEnforceWithMultipleRulesets()
     {
         $countries = include('tests/fixtures/countries.php');
         $visas = include('tests/fixtures/visas.php');
         $users = include('tests/fixtures/users.php');
 
-        foreach ($this->abac_array_a as $abac) {
-
-            // for this test, the attribute in error are the attributes tester une the last rule of the ruleset
-            $this->assertEquals(['age','code-iso-du-pays'], $abac->enforce('gunlaw', $users[2]));
+        foreach ($this->multipleRulesetSet as $abac) {
+            // for this test, the attribute in error are the tested attributes of the last rule of the ruleset
+            $this->assertFalse($abac->enforce('gunlaw', $users[2]));
+            $this->assertEquals(['age','code-iso-du-pays'], $abac->getErrors());
 
             $this->assertTrue($abac->enforce('gunlaw', $users[4]));
             $this->assertTrue($abac->enforce('gunlaw', $users[0]));
@@ -93,30 +83,28 @@ class AbacTest extends \PHPUnit_Framework_TestCase
     }
     
     
-    public function testEnforceGetterParams()
+    public function testEnforceWithGetterParams()
     {
         $countries = include('tests/fixtures/countries.php');
         $visas = include('tests/fixtures/visas.php');
         $users = include('tests/fixtures/users.php');
     
-        foreach ($this->abac_getter_params_a as $abac) {
-        
-            // for this test, the attribute in error are the attributes tester une the last rule of the ruleset
-            $this->assertEquals(['visa-specific'], $abac->enforce('travel-to-foreign-country', $users[0], $countries[2]));
+        foreach ($this->getterParamsSet as $abac) {
+            $this->assertFalse($abac->enforce('travel-to-foreign-country', $users[0], $countries[2]));
+            $this->assertEquals(['visa-specific'], $abac->getErrors());
             $this->assertTrue($abac->enforce('travel-to-foreign-country', $users[1], $countries[2]));
         }
     }
 
     
-    public function testEnforceImport()
+    public function testEnforceWithImport()
     {
         $countries = include('tests/fixtures/countries.php');
         $visas = include('tests/fixtures/visas.php');
         $users = include('tests/fixtures/users.php');
-        foreach ($this->abac_import_a as $abac) {
-            
-            // for this test, the attribute in error are the attributes tester une the last rule of the ruleset
-            $this->assertEquals(['visa-specific'], $abac->enforce('travel-to-foreign-country', $users[0], $countries[2]));
+        foreach ($this->importSet as $abac) {
+            $this->assertFalse($abac->enforce('travel-to-foreign-country', $users[0], $countries[2]));
+            $this->assertEquals(['visa-specific'], $abac->getErrors());
             $this->assertTrue($abac->enforce('travel-to-foreign-country', $users[1], $countries[2]));
         }
     }
