@@ -4,9 +4,9 @@ namespace PhpAbac\Test\Cache\Item;
 
 use PhpAbac\Cache\Item\MemoryCacheItem;
 
-class MemoryCacheItemTest extends \PHPUnit_Framework_TestCase
+class MemoryCacheItemTest extends \PHPUnit\Framework\TestCase
 {
-    /** @var \PhpAbac\Cache\Item\MemoryCacheItem **/
+    /** @var MemoryCacheItem **/
     protected $item;
     
     public function setUp()
