@@ -4,9 +4,9 @@ namespace PhpAbac\Test\Cache\Item;
 
 use PhpAbac\Cache\Item\TextCacheItem;
 
-class TextCacheItemTest extends \PHPUnit_Framework_TestCase
+class TextCacheItemTest extends \PHPUnit\Framework\TestCase
 {
-    /** @var \PhpAbac\Cache\Item\TextCacheItem **/
+    /** @var TextCacheItem **/
     protected $item;
 
     public function setUp()
