@@ -6,17 +6,18 @@ use PhpAbac\Comparison\ObjectComparison;
 
 use Symfony\Component\Config\FileLocator;
 
-use PhpAbac\Manager\ConfigurationManager;
-use PhpAbac\Manager\ComparisonManager;
-use PhpAbac\Manager\AttributeManager;
+use PhpAbac\Manager\{
+    AttributeManager,
+    ComparisonManager,
+    ConfigurationManager
+};
 
-use PhpAbac\Model\PolicyRuleAttribute;
-use PhpAbac\Model\Attribute;
+use PhpAbac\Example\{
+    User,
+    Vehicle
+};
 
-use PhpAbac\Example\User;
-use PhpAbac\Example\Vehicle;
-
-class ObjectComparisonTest extends \PHPUnit_Framework_TestCase
+class ObjectComparisonTest extends \PHPUnit\Framework\TestCase
 {
     /** @var ArrayComparison **/
     protected $comparison;
