@@ -4,12 +4,17 @@ namespace PhpAbac\Test\Manager;
 
 use Symfony\Component\Config\FileLocator;
 
-use PhpAbac\Manager\AttributeManager;
-use PhpAbac\Manager\ConfigurationManager;
-
+use PhpAbac\Manager\{
+    AttributeManager,
+    ConfigurationManager
+};
+use PhpAbac\Model\{
+    Attribute,
+    EnvironmentAttribute
+};
 use PhpAbac\Example\User;
 
-class AttributeManagerTest extends \PHPUnit_Framework_TestCase
+class AttributeManagerTest extends \PHPUnit\Framework\TestCase
 {
     /** @var \PhpAbac\Manager\AttributeManager **/
     private $manager;
@@ -26,7 +31,7 @@ class AttributeManagerTest extends \PHPUnit_Framework_TestCase
     {
         $attribute = $this->manager->getAttribute('main_user.age');
         
-        $this->assertInstanceOf('PhpAbac\Model\Attribute', $attribute);
+        $this->assertInstanceOf(Attribute::class, $attribute);
         $this->assertEquals('user', $attribute->getType());
         $this->assertEquals('Age', $attribute->getName());
         $this->assertEquals('age', $attribute->getSlug());
@@ -38,7 +43,7 @@ class AttributeManagerTest extends \PHPUnit_Framework_TestCase
     {
         $attribute = $this->manager->getAttribute('environment.service_state');
         
-        $this->assertInstanceOf('PhpAbac\Model\EnvironmentAttribute', $attribute);
+        $this->assertInstanceOf(EnvironmentAttribute::class, $attribute);
         $this->assertEquals('environment', $attribute->getType());
         $this->assertEquals('SERVICE_STATE', $attribute->getVariableName());
         $this->assertEquals('Statut du service', $attribute->getName());
