@@ -2,10 +2,12 @@
 
 namespace PhpAbac\Test\Model;
 
-use PhpAbac\Model\Attribute;
-use PhpAbac\Model\PolicyRuleAttribute;
+use PhpAbac\Model\{
+    Attribute,
+    PolicyRuleAttribute
+};
 
-class PolicyRuleAttributeTest extends \PHPUnit_Framework_TestCase
+class PolicyRuleAttributeTest extends \PHPUnit\Framework\TestCase
 {
     public function testEntity()
     {
