@@ -2,10 +2,12 @@
 
 namespace PhpAbac\Test\Model;
 
-use PhpAbac\Model\PolicyRule;
-use PhpAbac\Model\PolicyRuleAttribute;
+use PhpAbac\Model\{
+    PolicyRule,
+    PolicyRuleAttribute
+};
 
-class PolicyRuleTest extends \PHPUnit_Framework_TestCase
+class PolicyRuleTest extends \PHPUnit\Framework\TestCase
 {
     public function testEntity()
     {
