@@ -2,38 +2,36 @@
 
 namespace PhpAbac;
 
-use PhpAbac\Manager\AttributeManager;
-use PhpAbac\Manager\PolicyRuleManager;
-use PhpAbac\Manager\ConfigurationManager;
-use PhpAbac\Manager\CacheManager;
-use PhpAbac\Manager\ComparisonManager;
+use PhpAbac\Manager\{
+    AttributeManager,
+    CacheManager,
+    ComparisonManager,
+    ConfigurationManager,
+    PolicyRuleManager
+};
+use PhpAbac\Model\PolicyRuleAttribute;
 
 use Symfony\Component\Config\FileLocator;
 
-use PhpAbac\Model\PolicyRuleAttribute;
 
-class Abac
+final class Abac
 {
-    /** @var \PhpAbac\Manager\ConfigurationManager * */
+    /** @var ConfigurationManager **/
     private $configuration;
-    /** @var \PhpAbac\Manager\PolicyRuleManager * */
+    /** @var PolicyRuleManager **/
     private $policyRuleManager;
-    /** @var \PhpAbac\Manager\AttributeManager * */
+    /** @var AttributeManager **/
     private $attributeManager;
-    /** @var \PhpAbac\Manager\CacheManager * */
+    /** @var CacheManager **/
     private $cacheManager;
-    /** @var \PhpAbac\Manager\ComparisonManager * */
+    /** @var ComparisonManager **/
     private $comparisonManager;
+    /** @var array **/
+    private $errors;
     
-    /**
-     * @param array  $configPaths
-     * @param array  $cacheOptions     Option for cache
-     * @param string $configPaths_root The origin folder to find $configPaths
-     * @param array  $options
-     */
-    public function __construct($configPaths, $cacheOptions = [], $configPaths_root = null, $options = [])
+    public function __construct(array $configFiles, array $cacheOptions = [], string $configDir = null, array $options = [])
     {
-        $this->configure($configPaths, $configPaths_root);
+        $this->configure($configFiles, $configDir);
         $this->attributeManager = new AttributeManager($this->configuration->getAttributes(), $options);
         $this->policyRuleManager = new PolicyRuleManager($this->attributeManager, $this->configuration->getRules());
         $this->cacheManager      = new CacheManager($cacheOptions);
@@ -41,18 +39,13 @@ class Abac
     }
     
     /**
-     * @param array  $configPaths
-     * @param string $configPaths_root The origin folder to find $configPaths
+     * Read the given files contained in $configDir
      */
-    public function configure($configPaths, $configPaths_root = null)
+    public function configure(array $configFiles, string $configDir = null)
     {
-        //		foreach ( $configPaths as &$configPath ) {
-//			$configPath = $configPaths_root . $configPath;
-//		}
-        $locator             = new FileLocator($configPaths_root);
-        $this->configuration = new ConfigurationManager($locator);
-        $this->configuration->setConfigPathRoot($configPaths_root);
-        $this->configuration->parseConfigurationFile($configPaths);
+        $this->configuration = new ConfigurationManager(new FileLocator($configDir));
+        $this->configuration->setConfigPathRoot($configDir);
+        $this->configuration->parseConfigurationFile($configFiles);
     }
     
     /**
@@ -69,16 +62,10 @@ class Abac
      *
      * Available cache drivers are :
      * * memory
-     *
-     * @param string $ruleName
-     * @param object $user
-     * @param object $resource
-     * @param array  $options
-     *
-     * @return boolean|array
      */
-    public function enforce($ruleName, $user, $resource = null, $options = [])
+    public function enforce(string $ruleName, $user, $resource = null, array $options = []): bool
     {
+        $this->errors = [];
         // If there is dynamic attributes, we pass them to the comparison manager
         // When a comparison will be performed, the passed values will be retrieved and used
         if (isset($options[ 'dynamic_attributes' ])) {
@@ -92,17 +79,15 @@ class Abac
                 return $cacheValue;
             }
         }
-        $policyRule_a = $this->policyRuleManager->getRule($ruleName, $user, $resource);
+        $policyRules = $this->policyRuleManager->getRule($ruleName, $user, $resource);
         
-        foreach ($policyRule_a as $policyRule) {
+        foreach ($policyRules as $policyRule) {
             // For each policy rule attribute, we retrieve the attribute value and proceed configured extra data
             foreach ($policyRule->getPolicyRuleAttributes() as $pra) {
                 /** @var PolicyRuleAttribute $pra */
                 $attribute = $pra->getAttribute();
                 
                 $getter_params = $this->prepareGetterParams($pra->getGetterParams(), $user, $resource);
-//				var_dump($pra->getGetterParams());
-//				var_dump($getter_params);
                 $attribute->setValue($this->attributeManager->retrieveAttribute($attribute, $user, $resource, $getter_params));
                 if (count($pra->getExtraData()) > 0) {
                     $this->processExtraData($pra, $user, $resource);
@@ -111,17 +96,21 @@ class Abac
             }
             // The given result could be an array of rejected attributes or true
             // True means that the rule is correctly enforced for the given user and resource
-            $result = $this->comparisonManager->getResult();
-            if (true === $result) {
+            $this->errors = $this->comparisonManager->getResult();
+            if (count($this->errors) === 0) {
                 break;
             }
         }
         if ($cacheResult) {
-            $cacheItem->set($result);
+            $cacheItem->set((count($this->errors) > 0) ? $this->errors : true);
             $this->cacheManager->save($cacheItem);
         }
-        
-        return $result;
+        return count($this->errors) === 0;
+    }
+    
+    public function getErrors(): array
+    {
+        return $this->errors;
     }
     
     /**
@@ -151,12 +140,7 @@ class Abac
         return $values;
     }
     
-    /**
-     * @param \PhpAbac\Model\PolicyRuleAttribute $pra
-     * @param object                             $user
-     * @param object                             $resource
-     */
-    public function processExtraData(PolicyRuleAttribute $pra, $user, $resource)
+    private function processExtraData(PolicyRuleAttribute $pra, $user, $resource)
     {
         foreach ($pra->getExtraData() as $key => $data) {
             switch ($key) {
@@ -167,7 +151,6 @@ class Abac
                     // The "with" extra data is an array of attributes, which are objects
                     // Once we process it as policy rule attributes, we set it as the main policy rule attribute value
                     $subPolicyRuleAttributes = [];
-                    $extraData               = [];
                     
                     foreach ($this->policyRuleManager->processRuleAttributes($data, $user, $resource) as $subPolicyRuleAttribute) {
                         $subPolicyRuleAttributes[] = $subPolicyRuleAttribute;
