@@ -2,17 +2,9 @@
 
 namespace PhpAbac\Comparison;
 
-use PhpAbac\Model\Attribute;
-
 class ObjectComparison extends AbstractComparison
 {
-    /**
-     * @param string $attributeId
-     * @param mixed $value
-     * @param array $extraData
-     * @return boolean
-     */
-    public function isFieldEqual($attributeId, $value, $extraData = [])
+    public function isFieldEqual(string $attributeId, $value, array $extraData = []): bool
     {
         $attributeManager = $this->comparisonManager->getAttributeManager();
         // Create an attribute out of the extra data we have and compare its retrieved value to the expected one
