@@ -4,24 +4,12 @@ namespace PhpAbac\Comparison;
 
 class StringComparison extends AbstractComparison
 {
-    /**
-     * @param string $expected
-     * @param string $value
-     *
-     * @return bool
-     */
-    public function isEqual($expected, $value)
+    public function isEqual(string $expected, $value): bool
     {
         return $expected === $value;
     }
 
-    /**
-     * @param string $expected
-     * @param string $value
-     *
-     * @return bool
-     */
-    public function isNotEqual($expected, $value)
+    public function isNotEqual(string $expected, $value): bool
     {
         return !$this->isEqual($expected, $value);
     }
