@@ -1,31 +1,21 @@
 <?php
-/**
- * Projet :  php-abac.
- * User: mvedie
- * Date: 24/11/2016
- * Time: 18:51
- */
 
 namespace PhpAbac\Loader;
 
-use PhpAbac\Manager\ConfigurationManager;
-use Symfony\Component\Config\FileLocatorInterface;
 use Symfony\Component\Config\Loader\FileLoader;
 
 abstract class AbacLoader extends FileLoader
 {
     /**
-     * Must be overrided to contains an array of allowed extension
+     * @var array
+     *      Must be defined in child classes to contains an array of allowed extension
      */
-    protected static $_EXTENSION_ALLOWED_A = [];
-    
-    /** @var  ConfigurationManager The configuration manage instanciator and user of this AbacLoader Instance */
+    protected static $allowedExtensions = [];
+    /**
+     * @var ConfigurationManager
+     *      The configuration manage instanciator and user of this AbacLoader Instance
+     */
     protected $configurationManger;
-    
-    public function __construct(FileLocatorInterface $locator)
-    {
-        parent::__construct($locator);
-    }
     
     /**
      * Method to load a resource and return an array that contains decoded data of the resource
@@ -37,16 +27,6 @@ abstract class AbacLoader extends FileLoader
      */
     abstract public function load($resource, $type = null);
     
-//	/**
-//	 * Method to check if a resource ( by his path ) is supported by the loader
-//	 *
-//	 * @param string $resource The path of the resource to check
-//	 * @param null   $type     ??
-//	 *
-//	 * @return boolean Return true if the resource is supported by the loader
-//	 */
-//	abstract public function supports( $resource, $type = null );
-    
     /**
      * Method to check if a resource is supported by the loader
      * This method check only extension file
@@ -55,17 +35,17 @@ abstract class AbacLoader extends FileLoader
      *
      * @return boolean Return true if the extension of the ressource is supported by the loader
      */
-    final public static function supportsExtension($resource)
+    final public static function supportsExtension($resource): bool
     {
-        return in_array(pathinfo($resource, PATHINFO_EXTENSION), self::getExtensionAllowed());
+        return in_array(pathinfo($resource, PATHINFO_EXTENSION), self::getAllowedExtensions());
     }
 
     /**
      * Method to return allowed extension for file to load with the loader
      * @return mixed
      */
-    final private static function getExtensionAllowed()
+    final private static function getAllowedExtensions()
     {
-        return static::$_EXTENSION_ALLOWED_A;
+        return static::$allowedExtensions;
     }
 }
